pyproject.toml
[tool.poetry]
name = "repl-nix-projectinsight"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.45.0"
flask = "^3.0.3"
openpyxl = "^3.1.5"
python-docx = "^1.1.2"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
